+++ sample/index.ts
@@ -2,6 +2,10 @@
 // Order Flower Sample
 
 import * as Lex from 'lex-sdk';
+//let AWS = require('aws-sdk');
+
+//AWS.config.credentials = new AWS.SharedIniFileCredentials({profile: 'developer'});
+//AWS.config.update({region:'ap-northeast-1'});
 
 let bot: Lex.Bot;
 exports.handler = async function (event: Lex.IntentRequest, context: any) {
@@ -11,6 +15,11 @@ exports.handler = async function (event: Lex.IntentRequest, context: any) {
             .addRequestHandlers(
                 OrderIntentHandler)
             .addErrorHandler(ErrorHandler)
+            .withTableName("lex-sdk-sample","lex-sample")
+            .withAutoCreateTable(true)
+            // .withDynamoDbClient(
+            //     new AWS.DynamoDB({ apiVersion: "latest", region: "us-east-1" })
+            // )
             .create();
     }
     return bot.invoke(event, context);
@@ -24,7 +33,6 @@ const ErrorHandler = {
         const message =  { contentType: Lex.ContentType.PlainText, content: "ERROR " + error.message };
         return h.responseBuilder
         .getCloseResponse(
-            h.attributes,
             Lex.FulfillmentState.Fulfilled,
             message)
     }
@@ -34,7 +42,17 @@ const OrderIntentHandler: Lex.RequestHandler = {
     canHandle(h: Lex.HandlerInput) {
         return (h.intentName == 'OrderFlowers')
     },
-    handle(h: Lex.HandlerInput) {
+    async handle(h: Lex.HandlerInput) {
+
+        // セッション情報の取得
+        let attributes = await h.attributesManager.getPersistentAttributes();
+        if(attributes.count == undefined){
+            attributes.count = 0;
+        };
+        console.log("😍" + attributes.count);
+        attributes.count = Number(attributes.count) + 1;
+        await h.attributesManager.setPersistentAttributes(attributes);
+        await h.attributesManager.savePersistentAttributes();
 
         const flowerType = h.slots['FlowerType'];
         const date = h.slots['PickupDate'];
@@ -48,17 +66,20 @@ const OrderIntentHandler: Lex.RequestHandler = {
                     message = {contentType:Lex.ContentType.PlainText, content: validationResult.message};
                 }
                 return h.responseBuilder.getElicitSlotResponse(
-                    h.attributes,
                     h.intentName,
                     h.slots,
                     validationResult.violatedSlot,
                     message);
             }
             if (flowerType) {
-                h.attributes.Price = String(flowerType.length * 5); // Elegant pricing model
+                // let attributes = await h.attributesManager.getPersistentAttributes();
+                // console.log("attributes" + JSON.stringify(attributes));
+                // attributes.Price = String(flowerType.length * 5); // Elegant pricing model
+                // console.log("attributes" + JSON.stringify(attributes));
+                // h.attributesManager.setPersistentAttributes(attributes);
             }
             return h.responseBuilder
-                .getDelegateResponse(h.attributes, h.slots)
+                .getDelegateResponse(h.slots)
 
         } else {  // FulfillmentCodeHook
 
@@ -66,7 +87,6 @@ const OrderIntentHandler: Lex.RequestHandler = {
 
             return h.responseBuilder
             .getCloseResponse(
-                h.attributes,
                 Lex.FulfillmentState.Fulfilled,
                 message)
         }