+++ history.md
@@ -5,9 +5,23 @@
     * <tt>CardContentType</tt>を追加
     * <tt>ResponseCard</tt>の型定義を修正
 
-* 2019/04/03 Ver1.1.2
+* 2019/04/03 Ver0.1.2
     * <tt>RequestInterceptor</tt>の型定義を修正
     * <tt>ResponseInterceptor</tt>の型定義を修正
     * <tt>addRequestInterceptors()</tt>を追加
     * <tt>addResponseInterceptors()</tt>を追加
     * <tt>addErrorHandlers()</tt>を追加
+
+* 2019/04/25 Ver0.2.0
+    * <tt>withTableName()</tt>を追加
+    * <tt>withAutoCreateTable()</tt>を追加
+    * <tt>withDynamoDbClient()</tt>を追加
+    * <tt>AttributesManager</tt>を追加
+    * <tt>AttributesManager.getRequestAttributes()</tt>リクエスト内情報の取得
+    * <tt>AttributesManager.setRequestAttributes()</tt>リクエスト内情報の設定
+    * <tt>AttributesManager.getSessionAttributes()</tt>セッション情報の取得
+    * <tt>AttributesManager.setSessionAttributes()</tt>セッション情報の設定
+    * <tt>AttributesManager.getPersistentAttributes()</tt>永続化情報の取得
+    * <tt>AttributesManager.setPersistentAttributes()</tt>永続化情報の設定
+    
+